packages/sage-benchmark/pyproject.toml
[build-system]
requires = [
    "setuptools>=64",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "isage-benchmark"
dynamic = ["version"]
description = "SAGE Benchmark - RAG and experimental benchmarks for SAGE framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "IntelliStream Team", email = "shuhao_zhang@hust.edu.cn" },
]
keywords = [
    "sage",
    "benchmark",
    "rag",
    "retrieval-augmented-generation",
    "evaluation",
    "experiments",
    "intellistream",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

# Core dependencies - 只包含外部依赖
dependencies = [
    # Note: sage.data is included as a submodule in src/sage/data
    # and will be installed as part of this package

    # RAG and benchmark dependencies
    "pyyaml>=6.0",
    "chromadb>=0.4.0",
    "pymilvus>=2.3.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

license = {text = "MIT"}

[project.optional-dependencies]
# SAGE 内部依赖
sage-deps = [
    "isage-common>=0.1.0",
    "isage-kernel>=0.1.0",
    "isage-middleware>=0.1.0",
    "isage-libs>=0.1.0",
]

# Agent training benchmark dependencies
agent-training = [
    "isage-libs>=0.1.0",  # 核心依赖已包含 peft, accelerate, tensorboard, trl
    "isage-libs[finetune-gpu]>=0.1.0",  # GPU 加速（bitsandbytes, wandb）
    "matplotlib>=3.7.0",  # for generating comparison charts
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff==0.14.6",  # Pinned to match tools/pre-commit-config.yaml
]

all = [
    "isage-benchmark[sage-deps,dev,agent-training]",
]

[project.urls]
Homepage = "https://github.com/intellistream/SAGE"
Documentation = "https://intellistream.github.io/SAGE"
Repository = "https://github.com/intellistream/SAGE"
Issues = "https://github.com/intellistream/SAGE/issues"

[project.scripts]
# Control Plane Benchmark CLI
sage-cp-bench = "sage.benchmark.benchmark_control_plane.cli:main"
# Refiner Benchmark CLI
sage-refiner-bench = "sage.benchmark.benchmark_refiner.cli:main"

[tool.setuptools]
# Automatically discover all packages under src/
# This will find all directories with __init__.py files
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["sage*"]
namespaces = true

[tool.setuptools.dynamic]
version = {attr = "sage.benchmark._version.__version__"}

[tool.setuptools.package-data]
"sage.benchmark.benchmark_rag" = [
    "config/*.yaml",
    "data/**/*",
    "evaluation/config/*.yaml",
]

# ============================================================================
# Code Quality Configuration
# Extends from root ruff.toml for unified standards across all packages
# ============================================================================
[tool.ruff]
extend = "../../tools/ruff.toml"

[tool.mypy]
cache_dir = "../../.sage/cache/mypy"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
    "src",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
addopts = [
    "--benchmark-storage=../../.sage/benchmarks",
    "-o",
    "cache_dir=../../.sage/cache/pytest",
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "benchmark: marks tests as benchmark tests",
    "rag: marks tests as RAG benchmark tests",
]

[tool.coverage.run]
source = [
    "src/sage",
]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/_test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.black]
line-length = 100
target-version = ['py310']
